Add an optimization for removing redundant suspend tests in ART
This CL:
(1) eliminates redundant suspend checks (dominated by another check),

(2) removes the special treatment of the R4 register, which got reset on every
native call, possibly yielding long execution sequences without any suspend checks, and

(3) fixes the absence of suspend checks in leaf methods.

(2) and (3) increase the frequency of suspend checks, which improves the 
performance of GC and the accuracy of profile data. To compensate for the 
increased number of checks, we implemented an optimization that leverages 
dominance information to remove redundant suspend checks on back edges.  
Based on the results of running the Caffeine benchmark on Nexus 7, the patch
performs roughly 30% more useful suspend checks, spreading them much more evenly 
along the execution trace, while incurring less than 1% overhead.  For 
flexibility consideration, this CL defines two flags to control the enabling of
optimizations.  The original implementation is the default.

Change-Id: I31e81a5b3c53030444dbe0434157274c9ab8640f
Signed-off-by: Wei Jin <wejin@google.com>