[WIP] Arm64 hard-float
Very much a work in progress.  Blindly enabling most opcodes to
enable invoke and entry code testing.  Much of the hard-float code
should be able to be shared across 64 bit targets (and later, when
we roll to 32-bit hard-float, all targets).

However, we'll defer consolation to a later CL.

In a much later CL, hard-float support will be added for 32-bit
targets.

Change-Id: Ib66209285cad8998d77a14781de300af02a96b15