[WIP] Register promotion support for 64-bit targets
Not ready for review.

A significant amount of refactoring needs to be done, though
I may decide to push some of it into a later CL.  Also, we
may be able to avoid changing the vreg mappings at this time.
They will need to change if we are to support promotion into
arm32's d16..d31 - but perhaps that can wait.

Also required for reusing the current vmap scheme is ensuring
that a 32 and 64-bit views of the same vreg share the same
physical register.  I'd originally not planned on making this
restriction (which conflicts with arm32's d16..d31), but the
verifier should prevent any code patterns which would break
this for our other targets.

For example, if v4, v5, v4/v5 and v5/v6 are all hot enough to
promote, we'd end up with something like:

v4 (as an int)    -> r10
v4/v5 (as a long) -> r10
v5 (as an int)    -> r11
v5/v6 (as a long) -> r11

Applying this restriction allows us to delay adding a tag bit
in the vmap table until we need to fully support the extra
Thumb2 unaliased double precision registers.

Change-Id: I6a152b9c164d9f1a053622266e165428045362f3