ART: Compacting ROS spaces with semispace copy GC
Current semispace copy GC is mainly associated with bump pointer
spaces. Though it squeezes fragmentation most aggressively, an extra
copy is required to re-establish the data in the ROS space to allow
CMS GCs to happen afterwards. As semispace copy GC is still stop-the-world,
this not only introduces unnecessary overheads but also potential janks.
Using semispace copy GC to compact the data in a ROS space to another ROS
solves this problem. Although it squeezes less fragmentation, CMS GCs can
run immediately after the compaction. Compared with pure CMS GC, this ROS-to-ROS
copy not only reduces fragmentation but also frees physical pages.

We apply this algorithm in two cases:
1) right before throwing an OOM.
2) when app is switched to background.
For case 1), OOMs are significantly delayed in the harmony GC stress test,
with compaction ratio up to 0.87. For case 2), compaction ratio around 0.5 is
observed in both built-in SMS and browser. Similar results have been obtained
on other apps as well.

Change-Id: Iad9eabc6d046659fda3535ae20f21bc31f89ded3
Signed-off-by: Wang, Zuo <zuo.wang@intel.com>
Signed-off-by: Chang, Yang <yang.chang@intel.com>
Signed-off-by: Lei Li <lei.l.li@intel.com>
Signed-off-by: Lin Zang <lin.zang@intel.com>