[WIP] Add missing class initialization during compilation and tests
Adds missing class initialization during compilation and tests, especially
java.lang.Class. Otherwise, we'd be able to execute code while the referring
class is not initialized or initializing.

Adds class initialization checks in the interpreter. When entering the
interpreter: the class must either be initialized or be initializing by the
current thread (other threads must be waiting for the class initialization to
complete).

TODO: only enable AssertInitializedOrInitializingInThread in debug build.

Bug: 15899971
Change-Id: I9a4edd3739a3ca4cf1c4929dcbb44cdf7a1ca1fe