[WIP] Enable optimizations for Arm64
This CL enables optimizations that I believe are safe to enable.

The ones not enabled:

kLoadHoisting requires rework of PC def/use.  In arm32 we used updates
to the pc register to provide a barrier to hoisting loads above
branches.

kSuppressLoads requires kTrackLiveTemps, which is on hold while we
are working through the RefReg consistency issues.

kTrackLiveTemps needs debugging (likely problems around equality of
32 and 64-bit views of core registers (in particularly, those used
as references).

kPromoteRegs needs CL 99080 debugged and submitted.

kBBOpt may be ready, but will likely need debugging of the fused
branches.

Also, note that intrinsics and inlined methods are still turned off.

Change-Id: I20e4765925bebc0ac3b3001d199d403ed5a85fa6