Fix invalid elimination of load instructions.
https://android-review.googlesource.com/#/c/94205/ eliminate unnecessary
load in code generation stage, but it also slightly change the behavior
of EvalLoc(). Now the temp register can be live but hold a value can not
reflect the s-register.

If we want to avoid temp register go into this state, we will have to
make sure LoadValue must happen before EvalLoc.

This patch adds a new variable to track this state and remove the
limitation.

Change-Id: I61a29f0b0d1c625ecd52a71d71290bc8807365f3