Temporary fix for identifier size in the heap dump.
We're using addresses as IDs and previously we would have written
sizeof(void*) as the identifier size in the header. Although on 64 bits
this would be 8 bytes we kept writing only 4 bytes for the actual IDs
(only the low mem part of the address).

Writing the full address is the right thing to do here but it doesn't
work with other heap reading tools (after using hprof-conv). More
investigation is needed but in the meantime this is a good work around.

Bug: 16016862
Change-Id: I785abf5030681b6009370b7b2d7552e8b4b54e67