Use a special basic block evaluation order for GVN.
Use the topological sort order and return back to the start
of the loop if it needs recalculation when we reach its end.
Avoid merging obsolete LVNs through back branches, for
example when recalculating an inner loop for changes in the
outer loop, we behave as if we recalculated the inner loop
from scratch.

Also fix LocalValueNumbering::InPlaceIntersectMaps() which
was keeping only the last element of the intersection.

Change-Id: I4e10d4acb626a5b8a28ec0de106a7b37f9cbca32