Rewrite topological sort order and improve GVN.
Rewrite the topological sort order to include a full loop
before the blocks that go after the loop. Rewrite the
RepeatingTopologicalSortIterator to return to the loop head
if it needs recalculation when we reach the end of the loop.

In GVN, use the new repeating topological order iterator and
for a loop head merge only the preceding blocks' LVNs if
we're not currently recalculating this loop.

Also fix LocalValueNumbering::InPlaceIntersectMaps() which
was keeping only the last element of the intersection.

Change-Id: I4e10d4acb626a5b8a28ec0de106a7b37f9cbca32