AArch64: Fix def use and enable kLoadHoisting.
Unlike arm32, we are not able to write to pc register except branch
instruction, while branch instruction is marked as kEncodeAll. So it
would be safe to just ignore the use mask with PC.
With this trick, we do not need to introduce a fake resource mask for
pc register.

Change-Id: Ifc5a2484395486c01a64307a4acddc794026d46a