[wip] AArch64: Add Load Store Elimination for ARM64
This patch refactors the implementation of the LoadStoreElimination
optimisation pass. Please note that this pass was disabled and not
functional for any of the backends.

The current implementation tracks aliases and handles more DalvikRegs
and Heap memory regions. It has been tested and it is known to optimise
out the following:
  * LDR - LDR
  * STR - LDR
  * STR - STR

Change-Id: Iefae9b696f87f833ef35c451ed4d49c5a1b6fde0
Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>