Reduce stack usage for overflow checks
This reduces the stack space reserved for overflow checks to 8K, split
into a gap and a 4K protected region.

Also prevents signal runaway by detecting a signal inside code that
resulted from a signal handler invokation.  And adds a max signal count to
the SignalTest to prevent it running forever.

Bug: 15435566

Change-Id: Id4fd46f22d52d42a9eb431ca07948673e8fda694