[WIP] AArch64: experiments in ELF patching code.
For now this is just an hacky explorative patch. It does change oat
files patching in the following way:

1. It forces usage of "bl relative-offset" rather than
   "mov x30, literal-pool; blr x30" in some cases,

2. It loads class and methods via movz/movk rather than via the
   literal pool.

Note that this patch does not handle the case where the target of "bl"
is more than 128 MB distant from the actual "bl" instruction.
The patch currently aborts in such cases. Despite this, the patch
should offer a chance of measuring how much these changes could impact
the performance of Art.

Change-Id: I7b32dfb25d83b5218e2f3dd94080439346f58e86