Fix checks for kLiteral in local optimizations.
The check for kLiteral (literal load) just checked the kLiteral
bit in the def mask.  The kEncodeAll mask has the kLiteral bit
set so this check was triggering.  The fix is to check for
only the kLiteral bit being set and no other special bits.

The semantics of the special bits in the use/def mask is that
only one of them can be set at the same time.

Bug: 16824330

Change-Id: I0f1c1157e017870414ffef11767e5433d1fd4401