ART: Add clinit elimination for static invokes
When invoking a static method from a different class, we need to ensure
that class initialization is done for that class. However, we can guarantee
that clinit is done if a dominator block which contains static access to
same class exists. In those instances, we mark the MIR to know that class
is surely already initialized.

Change-Id: I755a00eb87df82cb2e264963731ccf5199585e8e
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>