Fix crash in VirtualMachine.AllThreads JDWP command
When collecting all the started threads, we may encounter a thread in the
process of being started from Java code (Thread.start) in the threads list. The
situation of the crash is this thread still points to its java.lang.Thread peer
through a JNI global reference (Thread::jpeer). When we want to access the peer
directly (Thred::opeer) by calling Thread::GetPeer method, we fail on a check
verifying the JNI global reference has been destroyed and reset.

To fix this, we only return threads whose Thread::jpeer has been deleted. This
indicates the thread is started and we're about to call its Thread.run method
(see Thread::CreateCallback for details).

Bug: 16660101

Change-Id: Ieef1aaf9c59a46dc5862e890a9f3ae0c9a4d98e0