Implement vmarko suspend checkss.
This is an optimization for the ARM and AARCH64 suspend checks
generated by the compiler.  It is based on an idea by vmarko
and uses a combination of a reserved register (r4 on ARM and
x19 on aarch64) and a single load instruction to perform
the suspend check.

It seems to give a 2% reduction in boot.oat and from a
rudimentary performance test, a 2% speed improvement.

A full performance run is required to verify these numbers.

Change-Id: If35ac79850dc3cea47a2d99f852c762d62f31d3e